core-engine/requirements.txt
requests>=2.31.0
psutil>=5.9.0
cryptography>=41.0.0
click>=8.1.0
colorama>=0.4.6
tqdm>=4.66.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
